feat: Enhance Quiz and Student services with long ID support and add Question mapping functionality

+++ 06-06-2025/QuizupAPI/Interfaces/IStudentService.cs
@@ -9,13 +9,13 @@ namespace QuizupAPI.Interfaces
     public interface IStudentService
     {
         Task<Student> AddStudentAsync(StudentAddRequestDTO student);
-        Task<Student> GetStudentByIdAsync(int id);
+        Task<Student> GetStudentByIdAsync(long id);
         Task<IEnumerable<Student>> GetAllStudentsAsync();
-        Task<Student> UpdateStudentAsync(int id, StudentUpdateRequestDTO student);
-        Task<Student> DeleteStudentAsync(int id);
-        Task<IEnumerable<QuizSubmission>> GetSubmissionsByStudentIdAsync(int id);
-        Task<QuizSubmission> SubmitQuizAsync(int studentId, QuizSubmissionDTO submission);
-        Task<QuizSubmission> SaveAnswersAsync(int studentId, QuizSubmissionDTO submission);
+        Task<Student> UpdateStudentAsync(long id, StudentUpdateRequestDTO student);
+        Task<Student> DeleteStudentAsync(long id);
+        Task<IEnumerable<QuizSubmission>> GetSubmissionsByStudentIdAsync(long id);
+        Task<QuizSubmission> SubmitQuizAsync(long studentId, QuizSubmissionDTO submission);
+        Task<QuizSubmission> SaveAnswersAsync(long studentId, QuizSubmissionDTO submission);
 
     }
 }
